+++ Localization/deDE.lua
@@ -1,4 +1,4 @@
--- [[ Exported at 2022-08-07 15-23-10 ]] --
+-- [[ Exported at 2022-08-11 18-16-17 ]] --
 -- [[ This code is automatically generated as an export from ]] --
 -- [[ an SQLite database and is not meant for manual edit. ]] --